webapp/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-pptx==0.6.23
playwright==1.41.0
requests==2.31.0
python-multipart==0.0.6
